FST-80 #comment Small refactoring #in-progress
